saving

+++ src/halc_allocators.h
@@ -1,6 +1,7 @@
 #ifndef _HALC_ALLOCATORS_H_
 #define _HALC_ALLOCATORS_H_
 
+#include <stdlib.h>
 #include "halc_errors.h"
 
 // ==================== Allocators ======================
@@ -9,7 +10,7 @@ errc setupDefaultAllocator();
 
 // These are the only two functions I'm going to support in my allocator interface.
 struct allocator{
-    void* (*malloc_fn) (unsigned long long);
+    void* (*malloc_fn) (size_t);
     void (*free_fn) (void*);
 };
 
@@ -19,11 +20,11 @@ extern struct allocator gDefaultAllocator;
 
 // allows you to bring whatever malloc and free function you want
 errc setupCustomDefaultAllocator(
-        void* (*malloc_fn) (usize),
+        void* (*malloc_fn) (size_t),
         void (*free_fn) (void*));
 
 // halcyon default allocation function
-errc halloc_advanced(void** ptr, usize size);
+errc halloc_advanced(void** ptr, size_t size);
 
 #define halloc(ptr, size) halloc_advanced((void**) ptr, size)
